refactor(painters/dom): co-locate image utils under images/
Move image-clip-path and image-selectors (plus their tests) from
utils/ into the images/ directory alongside image-block, drawing-image,
hyperlink, and image-fragment. Update import paths in callers. No
behavior change.
