+++ packages/layout-engine/painters/dom/src/images/image-clip-path.test.ts
