+++ packages/layout-engine/painters/dom/src/images/image-selectors.ts
